UCAI_MVP.V2.0/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9
google-generativeai>=0.5.0
google-cloud-firestore
google-cloud-bigquery
google-cloud-pubsub
google-cloud-storage
google-cloud-monitoring
google-cloud-asset
google-cloud-billing
google-api-python-client>=2.100.0
google-auth>=2.0.0
networkx>=3.2
pandas>=2.0.0
pydantic>=2.0.0
